fix theme switcher
