+++ dotcom/src/routes/__layout.svelte
@@ -19,27 +19,6 @@
 	import "../app.css";
 	import "../ibm-plex-var.css";
 
-	const setTheme = (theme: "light" | "dark" | "default") => {
-		document.body.classList.add("themed");
-		const { classList } = document.querySelector("html");
-		classList.remove("light", "dark", "default");
-		switch (theme) {
-			case "light":
-				classList.add("light");
-				localStorage.setItem("theme", "light");
-				break;
-			case "dark":
-				classList.add("dark");
-				localStorage.setItem("theme", "dark");
-				break;
-			case "default":
-				classList.add("default");
-				localStorage.removeItem("theme");
-				break;
-		}
-		return theme;
-	};
-
 	export let lang: Lang;
 </script>
 
@@ -48,10 +27,6 @@
 <main>
 	<slot />
 </main>
-
-<button on:click={() => setTheme("light")}>light</button>
-<button on:click={() => setTheme("dark")}>dark</button>
-<button on:click={() => setTheme("default")}>reset</button>
 
 <Footer {lang} />
